* merge bin and lib completely * convert some ancient RCSID styles etc. to more modern stuff * fix use of CDIAGFLAGS+=, CFLAGS+= and COPTS+= to the correct   variant (include bsd.own.mk _then_ COPTS+=) so they're added   after any user-specified flags   (this is probably lurking in more places than I can imagine) * fix some indenting, comments, etc. * make stuff compile (fix warnings/errors) * add script to review _all_ shlib_version after an openbsd import * remove two more obsolete /usr/lib/debug/ references * bin/ps: comment out some SMP code

+++ src/lib/libevent/kqueue.c
@@ -1,3 +1,4 @@
+/**	$MirOS$ */
 /*	$OpenBSD: kqueue.c,v 1.18 2005/07/02 07:15:13 grunk Exp $	*/
 
 /*
@@ -133,7 +134,7 @@ kq_init(void)
 	 */
 	if (kevent(kq,
 		kqueueop->changes, 1, kqueueop->events, NEVENT, NULL) != 1 ||
-	    kqueueop->events[0].ident != -1 ||
+	    kqueueop->events[0].ident != (u_int)-1 ||
 	    kqueueop->events[0].flags != EV_ERROR) {
 		event_warn("%s: detected broken kqueue; not using.", __func__);
 		free(kqueueop->changes);
